🚀 Nova: モジュールと構造の分離・リファクタリング (#385)
💡 何を:
- `opendaw-wasm/src/engine/mixer.rs` と `opendaw-wasm/src/engine/stream.rs` から将来のルーティングとバスの処理を分離するため、`routing.rs` と `bus.rs` モジュールを作成しました。
- `opendaw-wasm/src/state/track.rs` を分割し、エフェクトやシンセ設定を `opendaw-wasm/src/state/track_plugin.rs` に分離しました。また、将来のクリップ管理のために `track_clip.rs` を追加しました。
- `opendaw-wasm/src/engine/synth.rs` のオシレーター関連処理を `oscillator.rs` に、ADSRボイス管理処理を `voice.rs` に分離しました。既存のロジックが壊れないよう再エクスポートを行いました。

🎯 なぜ:
- ROADMAP.md のフェーズ 2 のタスク `[2-1]`, `[2-2]` (routingとbusの作成)、フェーズ 4 のタスク `[4-1]`, `[4-2]` (trackのドメイン分割)、およびフェーズ 5 のタスク `[5-1]`, `[5-2]` (synthのボイス・オシレーター分割) に対応し、モジュール構造を整理するため。

📁 変更ファイル:
- `opendaw-wasm/src/engine/mod.rs`
- `opendaw-wasm/src/engine/synth.rs`
- `opendaw-wasm/src/engine/oscillator.rs` (新規)
- `opendaw-wasm/src/engine/voice.rs` (新規)
- `opendaw-wasm/src/engine/routing.rs` (新規)
- `opendaw-wasm/src/engine/bus.rs` (新規)
- `opendaw-wasm/src/state/mod.rs`
- `opendaw-wasm/src/state/track.rs`
- `opendaw-wasm/src/state/track_plugin.rs` (新規)
- `opendaw-wasm/src/state/track_clip.rs` (新規)
- `opendaw-wasm/src/engine/mixer.rs`

✅ 検証:
- `cargo check --manifest-path=opendaw-wasm/Cargo.toml`
- `cargo test --manifest-path=opendaw-wasm/Cargo.toml`
すべてのテストがパスすることを確認しました。

PR created automatically by Jules

+++ opendaw-wasm/src/engine/mixer.rs
@@ -89,7 +89,7 @@ pub fn mix_tracks(out_buffer: &mut [f32], out_channels: u16, tracks: &mut [Track
             // オシレーターのサンプルを加算
             if let Some(osc) = track.oscillator.as_mut() {
                 if track.active_note_count > 0 {
-                    osc.set_frequency(track.active_notes[0]);
+                    osc.set_frequency(track.active_notes[0] as f32);
                     osc.set_active(true);
                 } else {
                     osc.set_active(false);
